Added signOut context method. Renamed folder context to hooks.
