+++ src/Pages/SignIn/index.tsx
@@ -4,7 +4,7 @@ import { Form } from '@unform/web';
 import { FormHandles } from '@unform/core';
 import * as Yup from 'yup';
 
-import { useAuth } from '../../context/AuthContext';
+import { useAuth } from '../../hooks/AuthContext';
 import logoImg from '../../assets/logoImg.svg';
 import getValidationErrors from '../../utils/getValidationErrors';
 import Input from '../../components/Input';
